requirements.txt
flask
dash
dash-bootstrap-components
openai
markitdown
pdf2image
pypandoc
python-magic